refactor: パッケージ整理のため、RepositoryItem を data.model 配下に移動した。
