+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/RepositoryItemListAdapter.kt
@@ -7,6 +7,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import jp.co.yumemi.android.code_check.data.model.RepositoryItem
 
 private val diffUtil = object : DiffUtil.ItemCallback<RepositoryItem>() {
     override fun areItemsTheSame(oldItem: RepositoryItem, newItem: RepositoryItem): Boolean {
