+++ app/src/main/res/navigation/nav_graph.xml
@@ -22,7 +22,7 @@
         tools:layout="@layout/fragment_repository_detail">
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.code_check.RepositoryItem" />
+            app:argType="jp.co.yumemi.android.code_check.data.model.RepositoryItem" />
     </fragment>
 
 </navigation>
